change data types from float based to double

+++ include/EvoAPI.hpp
@@ -7,8 +7,8 @@
 #include "XoshiroCpp.hpp"
 
 struct EvoDataSet {
-    Eigen::MatrixXf predictor;
-    Eigen::MatrixXf target;
+    Eigen::MatrixXd predictor;
+    Eigen::MatrixXd target;
 };
 
 class EvoAPI {
@@ -17,14 +17,14 @@ public:
     int generation_size_limit;
     int generation_count_limit;
 
-    Eigen::MatrixXf x;
-    Eigen::MatrixXf y;
+    Eigen::MatrixXd x;
+    Eigen::MatrixXd y;
 
     std::vector<std::vector<EvoIndividual>> population;
-    std::vector<float> fitness_history;
-    std::vector<float> titan_history;
+    std::vector<double> fitness_history;
+    std::vector<double> titan_history;
 
-    std::unordered_map<std::string, float> cache;
+    std::unordered_map<std::string, double> cache;
 
     EvoAPI(const std::string&);
     void setBoundaryConditions(unsigned int generation_size_limit, unsigned int generation_count_limit);
@@ -35,8 +35,8 @@ private:
     EvoIndividual titan;
 
     void setTitan(EvoIndividual, int);
-    void create_regression_input(std::tuple<int, int, std::vector<float>>);
+    void create_regression_input(std::tuple<int, int, std::vector<double>>);
     void append_generation_zero(XoshiroCpp::Xoshiro256Plus&);
 
-    EvoDataSet data_transformation_cacheless(Eigen::MatrixXf, Eigen::VectorXf, EvoIndividual&);
+    EvoDataSet data_transformation_cacheless(Eigen::MatrixXd, Eigen::VectorXd, EvoIndividual&);
 };
