+++ include/IOTools.hpp
@@ -3,4 +3,4 @@
 #include <iostream>
 #include <tuple>
 
-std::tuple<int, int, std::vector<float>> parse_csv(const std::string&);
+std::tuple<int, int, std::vector<double>> parse_csv(const std::string&);
